Move MultiTargetTerminationStrategy to test package
